requirements.txt
discord.py>=2.3.0
pymongo>=4.0.0
dnspython>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
requests>=2.28.0